requirements.txt
discord.py==2.3.2
boto3==1.34.0
pydantic==1.10.12
pyyaml==6.0.1
python-dotenv==1.0.0
